+++ CHANGELOG.rst
@@ -1,6 +1,14 @@
 Changelog
 ---------
 
+1.2.2 (2018-10-08)
+------------------
+* Change maximum retry counts for task `atol_receive_receipt_report`. Now its awaiting report for 29 hours.
+
+* Changed `atol_retry_created_receipts` and `atol_retry_initiated_receipts` tasks retry period.
+  Now it will retry receipts from day before yesterday
+
+
 1.2.1 (2018-05-22)
 ------------------
 * AtolAPI.base_url specifying in settings